Refactor the db init and migrations into a different package
