+++ api/cmd/main.go
@@ -2,7 +2,6 @@ package main
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"net/http"
 	"os"
@@ -11,16 +10,12 @@ import (
 
 	"github.com/gojek/mlp/api/pkg/authz/enforcer"
 	"github.com/gorilla/mux"
-	"github.com/golang-migrate/migrate/v4"
-	"github.com/golang-migrate/migrate/v4/database/postgres"
-	_ "github.com/golang-migrate/migrate/v4/source/file"
 	"github.com/heptiolabs/healthcheck"
-	"github.com/jinzhu/gorm"
-	_ "github.com/jinzhu/gorm/dialects/postgres"
 	"github.com/rs/cors"
 
 	"github.com/gojek/mlp/api/api"
 	"github.com/gojek/mlp/api/config"
+	"github.com/gojek/mlp/api/database"
 	"github.com/gojek/mlp/api/log"
 	"github.com/gojek/mlp/api/service"
 	"github.com/gojek/mlp/api/storage"
@@ -32,21 +27,12 @@ func main() {
 		log.Panicf("Failed initializing config: %v", err)
 	}
 
-	db, err := gorm.Open(
-		"postgres",
-		fmt.Sprintf("host=%s port=%d user=%s dbname=%s password=%s sslmode=disable",
-			cfg.DbConfig.Host,
-			cfg.DbConfig.Port,
-			cfg.DbConfig.User,
-			cfg.DbConfig.Database,
-			cfg.DbConfig.Password))
+	// init db
+	db, err := database.InitDB(&cfg.DbConfig)
 	if err != nil {
 		panic(err)
 	}
-	db.LogMode(false)
 	defer db.Close()
-
-	runDBMigration(db, cfg.DbConfig.MigrationPath)
 
 	applicationService, _ := service.NewApplicationService(db)
 	authEnforcer, _ := enforcer.NewEnforcerBuilder().
@@ -165,19 +151,4 @@ func (h uiHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 
 	// otherwise, use http.FileServer to serve the static dir
 	http.FileServer(http.Dir(h.staticPath)).ServeHTTP(w, r)
-}
-
-func runDBMigration(db *gorm.DB, migrationPath string) {
-	driver, err := postgres.WithInstance(db.DB(), &postgres.Config{})
-	if err != nil {
-		panic(err)
-	}
-
-	m, err := migrate.NewWithDatabaseInstance(migrationPath, "postgres", driver)
-	if err != nil {
-		panic(err)
-	}
-	if err := m.Up(); err != nil && !errors.Is(err, migrate.ErrNoChange) {
-		panic(err)
-	}
 }
